requirements.txt
djangorestframework_simplejwt
Django
requests
djangorestframework
pycryptodome
mysqlclient
django-cors-headers
django-filter